GEMAH - D3.00 - Materiels
- ADDED: Il est maintenant possible de voir les informations d'un matériel
- MODIFIED: Le test vérifie l'association avec un élève
- MODIFIED: La suppression vérifie que le matériel n'est pas associé à un élève
- MODIFIED: Les messages flashs sont maitenant mieux formatés

+++ app/Http/Controllers/Materiels/StockMaterielController.php
@@ -24,8 +24,9 @@ class StockMaterielController extends Controller
 		$etats = EtatMateriel::orderBy("nom")->get();
 
 		if ($request->exists(["type_id", "etat_id", "marque", "modele", "num_serie"])) {
-			$searchedMateriels = Materiel::search($request->input("type_id"), $request->input("etat_id"), $request->input("marque"), $request->input("modele"), $request->input("num_serie"))->get();
-		}  else {
+			$searchedMateriels = Materiel::search($request->input("type_id"), $request->input("etat_id"), $request->input("marque"), $request->input("modele"), $request->input("num_serie"))
+				->get();
+		} else {
 			$latestCreatedMateriels = Materiel::latestCreated()->take(10)->get();
 			$latestUpdatedMateriels = Materiel::latestUpdated()->take(10)->get();
 		}
@@ -71,7 +72,7 @@ class StockMaterielController extends Controller
 			"date_facture"          => "nullable",
 			"date_service_fait"     => "nullable",
 			"date_fin_garantie"     => "nullable",
-			"acheter_pour"          => "nullable",
+			"achat_pour"            => "nullable",
 		]);
 
 		Materiel::create($request->only([
@@ -90,7 +91,7 @@ class StockMaterielController extends Controller
 			"date_facture",
 			"date_service_fait",
 			"date_fin_garantie",
-			"acheter_pour",
+			"achat_pour",
 		]));
 
 		return redirect(route("web.materiels.stocks.index"));
@@ -100,11 +101,11 @@ class StockMaterielController extends Controller
 	 * Display the specified resource.
 	 *
 	 * @param Materiel $stock
-	 * @return void
+	 * @return View
 	 */
-	public function show(Materiel $stock)
+	public function show(Materiel $stock): View
 	{
-		//
+		return view("web.materiels.stocks.show", compact("stock"));
 	}
 
 	/**
@@ -147,7 +148,7 @@ class StockMaterielController extends Controller
 			"date_facture"          => "nullable",
 			"date_service_fait"     => "nullable",
 			"date_fin_garantie"     => "nullable",
-			"acheter_pour"          => "nullable",
+			"achat_pour"            => "nullable",
 		]);
 
 		$stock->update($request->only([
@@ -166,7 +167,7 @@ class StockMaterielController extends Controller
 			"date_facture",
 			"date_service_fait",
 			"date_fin_garantie",
-			"acheter_pour",
+			"achat_pour",
 		]));
 
 		return redirect(route("web.materiels.stocks.index"));
@@ -181,8 +182,12 @@ class StockMaterielController extends Controller
 	 */
 	public function destroy(Materiel $stock): RedirectResponse
 	{
-		$stock->delete();
+		if ($stock->eleve_id === null) {
+			$stock->delete();
 
-		return redirect(route("web.materiels.stocks.index"));
+			return redirect(route("web.materiels.stocks.index"));
+		}
+
+		return back()->withErrors("Impossible de supprimer un matériel lorsqu'il est associé à un élève");
 	}
 }
